LayoutTests/webgl/webgl2-init-missing-fragment-output-components-expected.txt

PASS [...pixel].join(',') is '255,0,0,0'
